requirements.txt
requests==2.22.0
keyboard==0.13.5
numpy==1.19.1
Pillow==7.2.0
PyQt5==5.15.0